+++ src/php/Gdbots/Schemas/Pbjx/Mixin/GetEventsRequest/GetEventsRequest.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Gdbots\Schemas\Pbjx\Mixin\GetEventsRequest;
+
+use Gdbots\Schemas\Pbjx\Mixin\Request\RequestV1;
+
+interface GetEventsRequest extends RequestV1
+{
+}